requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
crispy-bootstrap5==0.6
cryptography==38.0.0
defusedxml==0.7.1
Django==4.0.5
django-allauth==0.51.0
django-bootstrap-pagination==1.7.1
django-ckeditor==6.5.1
django-crispy-forms==1.14.0
django-etc==1.3.1
django-hitcount==1.3.5
django-js-asset==2.0.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
django-social-share==2.3.0
idna==3.3
mysqlclient==2.1.0
oauthlib==3.2.0
Pillow==9.1.1
pycparser==2.21
PyJWT==2.4.0
python3-openid==3.2.0
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
sqlparse==0.4.2
twilio==7.14.1
urllib3==1.26.12
whitenoise==6.2.0
